requirements.txt
ollama
openai~=1.65.2
phidata
rich
duckduckgo-search
